Make BlockHeader.Internal module; use regular fields in tests
