+++ chainweb.cabal
@@ -121,6 +121,7 @@ library
         , Chainweb.BlockCreationTime
         , Chainweb.BlockHash
         , Chainweb.BlockHeader
+        , Chainweb.BlockHeader.Internal
         , Chainweb.BlockHeader.Genesis.RecapDevelopment0Payload
         , Chainweb.BlockHeader.Genesis.RecapDevelopment1to9Payload
         , Chainweb.BlockHeader.Genesis.RecapDevelopment10to19Payload
