+++ test/Chainweb/Test/Mempool/Consensus.hs
@@ -8,7 +8,6 @@ module Chainweb.Test.Mempool.Consensus
   ( tests
   ) where
 
-import Control.Lens (view)
 import Control.Monad.IO.Class
 import Control.Monad.Trans.Resource
 
@@ -37,7 +36,7 @@ import Test.Tasty.QuickCheck
 
 import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
-import Chainweb.BlockHeader
+import Chainweb.BlockHeader.Internal
 import Chainweb.BlockHeaderDB
 import Chainweb.BlockHeaderDB.Internal (unsafeInsertBlockHeaderDb)
 import Chainweb.BlockWeight
@@ -347,18 +346,18 @@ header' h = do
         . newMerkleLog
         $ mkFeatureFlags
             :+: t'
-            :+: view blockHash h
+            :+: _blockHash h
             :+: target
             :+: _payloadWithOutputsPayloadHash (testBlockPayload h)
             :+: _chainId h
-            :+: BlockWeight (targetToDifficulty target) + view blockWeight h
-            :+: succ (view blockHeight h)
+            :+: BlockWeight (targetToDifficulty target) + _blockWeight h
+            :+: succ (_blockHeight h)
             :+: _versionCode v
             :+: epochStart (ParentHeader h) mempty t'
             :+: nonce
             :+: MerkleLogBody mempty
    where
-    BlockCreationTime t = view blockCreationTime h
+    BlockCreationTime t = _blockCreationTime h
     target = powTarget (ParentHeader h) mempty t'
     v = _chainwebVersion h
     t' = BlockCreationTime (scaleTimeSpan (10 :: Int) second `add` t)
@@ -434,12 +433,12 @@ instance Show ForkInfo where
         ++ debugTrans "newForkTrans" fiNewForkTrans
         ++ "\n\t"
         ++ "'head' of old fork:"
-        ++ "\n\t\tblock height: " ++ show (view blockHeight fiOldHeader)
-        ++ "\n\t\tblock hash: " ++ show (view blockHash fiOldHeader)
+        ++ "\n\t\tblock height: " ++ show (_blockHeight fiOldHeader)
+        ++ "\n\t\tblock hash: " ++ show (_blockHash fiOldHeader)
         ++ "\n\t"
         ++ "'head' of new fork:"
-        ++ "\n\t\tblock height: " ++ show (view blockHeight fiNewHeader)
-        ++ "\n\t\tblock hash: " ++ show (view blockHash fiNewHeader)
+        ++ "\n\t\tblock height: " ++ show (_blockHeight fiNewHeader)
+        ++ "\n\t\tblock hash: " ++ show (_blockHash fiNewHeader)
         ++ concatMap (debugHeader "main trunk headers") fiPreForkHeaders
         ++ concatMap (debugHeader "left fork headers") fiLeftForkHeaders
         ++ concatMap (debugHeader "right fork headers") fiRightForkHeaders
@@ -449,9 +448,9 @@ instance Show ForkInfo where
 debugHeader :: String -> BlockHeader -> String
 debugHeader context bh =
     "\nBlockheader from " ++ context ++ ": "
-    ++ "\n\t\tblockHeight: " ++ show (view blockHeight bh) ++ " (0-based)"
-    ++ "\n\t\tblockHash: " ++ show (view blockHash bh)
-    ++ "\n\t\tparentHash: " ++ show (view blockParent bh)
+    ++ "\n\t\tblockHeight: " ++ show (_blockHeight bh) ++ " (0-based)"
+    ++ "\n\t\tblockHash: " ++ show (_blockHash bh)
+    ++ "\n\t\tparentHash: " ++ show (_blockParent bh)
     ++ "\n"
 
 ----------------------------------------------------------------------------------------------------
