+++ test/Chainweb/Test/Pact/PactMultiChainTest.hs
@@ -51,7 +51,7 @@ import Pact.Types.SPV
 import Pact.Types.Term
 
 import Chainweb.BlockCreationTime
-import Chainweb.BlockHeader
+import Chainweb.BlockHeader.Internal
 import Chainweb.BlockHeight
 import Chainweb.ChainId
 import Chainweb.Cut
@@ -1560,7 +1560,7 @@ filterBlock f (MempoolBlock b) = MempoolBlock $ \mi ->
 
 blockForChain :: ChainId -> MempoolBlock -> MempoolBlock
 blockForChain chid = filterBlock $ \bh ->
-  view blockChainId bh == chid
+  _blockChainId bh == chid
 
 runCut' :: PactTestM ()
 runCut' = do
@@ -1682,7 +1682,7 @@ runToHeight :: BlockHeight -> PactTestM ()
 runToHeight bhi = do
   chid <- view menvChainId
   bh <- getHeader chid
-  when (view blockHeight bh < bhi) $ do
+  when (_blockHeight bh < bhi) $ do
     runCut'
     runToHeight bhi
 
@@ -1742,8 +1742,8 @@ signSender00 = set cbSigners [mkEd25519Signer' sender00 []]
 
 setFromHeader :: BlockHeader -> CmdBuilder -> CmdBuilder
 setFromHeader bh =
-  set cbChainId (view blockChainId bh)
-  . set cbCreationTime (toTxCreationTime $ _bct $ view blockCreationTime bh)
+  set cbChainId (_blockChainId bh)
+  . set cbCreationTime (toTxCreationTime $ _bct $ _blockCreationTime bh)
 
 buildBasic
     :: PactRPC T.Text
@@ -1795,7 +1795,7 @@ getPWO :: ChainId -> PactTestM (PayloadWithOutputs,BlockHeader)
 getPWO chid = do
   (TestBlockDb _ pdb _) <- view menvBdb
   h <- getHeader chid
-  Just pwo <- liftIO $ lookupPayloadWithHeight pdb (Just $ view blockHeight h) (view blockPayloadHash h)
+  Just pwo <- liftIO $ lookupPayloadWithHeight pdb (Just $ _blockHeight h) (_blockPayloadHash h)
   return (pwo,h)
 
 getHeader :: ChainId -> PactTestM BlockHeader
