+++ test/Chainweb/Test/Pact/PactReplay.hs
@@ -29,7 +29,7 @@ import Test.Tasty.HUnit
 -- chainweb imports
 
 import Chainweb.BlockCreationTime
-import Chainweb.BlockHeader
+import Chainweb.BlockHeader.Internal
 import Chainweb.BlockHeaderDB.Internal (unsafeInsertBlockHeaderDb)
 import Chainweb.Graph
 import Chainweb.Test.Cut.TestBlockDb
@@ -95,16 +95,16 @@ onRestart mpio iop step = do
     (_, _, bdb) <- iop
     bhdb' <- getBlockHeaderDb cid bdb
     block <- maxEntry bhdb'
-    step $ "max block has height " <> sshow (view blockHeight block)
-    let nonce = Nonce $ fromIntegral $ view blockHeight block
+    step $ "max block has height " <> sshow (_blockHeight block)
+    let nonce = Nonce $ fromIntegral $ _blockHeight block
     step "mine block on top of max block"
     T3 _ b _ <- mineBlock nonce iop
-    assertEqual "Invalid BlockHeight" 1 (view blockHeight b)
+    assertEqual "Invalid BlockHeight" 1 (_blockHeight b)
 
 testMemPoolAccess :: MemPoolAccess
 testMemPoolAccess = mempty
     { mpaGetBlock = \_g validate bh hash ph -> do
-        let (BlockCreationTime t) = view blockCreationTime ph
+        let (BlockCreationTime t) = _blockCreationTime ph
         getTestBlock t validate bh hash
     }
   where
@@ -181,7 +181,7 @@ serviceInitializationAfterFork mpio genesisBlock iop = do
     mineLine start ncounter len =
       evalStateT (mapM (const go) [startHeight :: Word64 .. (startHeight + len)]) start
         where
-          startHeight = fromIntegral $ view blockHeight start
+          startHeight = fromIntegral $ _blockHeight start
           go = do
               n <- liftIO $ Nonce <$> readIORef ncounter
               ret@(T3 _ newblock _) <- liftIO $ mineBlock n iop
@@ -229,7 +229,7 @@ firstPlayThrough mpio genesisBlock iop = do
     mineLine start ncounter len =
       evalStateT (mapM (const go) [startHeight :: Word64 .. (startHeight + len)]) start
         where
-          startHeight = fromIntegral $ view blockHeight start
+          startHeight = fromIntegral $ _blockHeight start
           go = do
               n <- liftIO $ Nonce <$> readIORef ncounter
               ret@(T3 _ newblock _) <- liftIO $ mineBlock n iop
@@ -280,12 +280,12 @@ testDeepForkLimit mpio (RewindLimit deepForkLimit) iop step = do
     let pdb = _bdbPayloadDb bdb
     step "query max db entry"
     maxblock <- maxEntry bhdb
-    pd <- lookupPayloadWithHeight pdb (Just $ view blockHeight maxblock) (view blockPayloadHash maxblock) >>= \case
+    pd <- lookupPayloadWithHeight pdb (Just $ _blockHeight maxblock) (_blockPayloadHash maxblock) >>= \case
       Nothing -> assertFailure "max block payload not found"
       Just x -> return x
     let maxblockPayload = payloadWithOutputsToPayloadData pd
-    step $ "max block has height " <> sshow (view blockHeight maxblock)
-    nonceCounterMain <- newIORef (fromIntegral $ view blockHeight maxblock)
+    step $ "max block has height " <> sshow (_blockHeight maxblock)
+    nonceCounterMain <- newIORef (fromIntegral $ _blockHeight maxblock)
 
     -- mine the main line a bit more
     step "mine (deepForkLimit + 1) many blocks on top of max block"
@@ -302,11 +302,11 @@ testDeepForkLimit mpio (RewindLimit deepForkLimit) iop step = do
     mineLine start ncounter len =
       evalStateT (mapM (const go) [startHeight :: Word64 .. (startHeight + len)]) start
         where
-          startHeight = fromIntegral $ view blockHeight start
+          startHeight = fromIntegral $ _blockHeight start
           go = do
               pblock <- gets ParentHeader
               n <- liftIO $ Nonce <$> readIORef ncounter
-              liftIO $ step $ "mine block on top of height " <> sshow (view blockHeight $ _parentHeader pblock)
+              liftIO $ step $ "mine block on top of height " <> sshow (_blockHeight $ _parentHeader pblock)
               ret@(T3 _ newblock _) <- liftIO $ mineBlock n iop
               liftIO $ modifyIORef' ncounter succ
               put newblock
@@ -331,7 +331,7 @@ mineBlock nonce iop = timeout 5000000 go >>= \case
       let
         creationTime = BlockCreationTime
           . add (TimeSpan 1_000_000)
-          . _bct . view blockCreationTime
+          . _bct . _blockCreationTime
           $ _parentHeader ph
 
       let bh = newBlockHeader
@@ -346,7 +346,7 @@ mineBlock nonce iop = timeout 5000000 go >>= \case
 
       (_, _, bdb) <- iop
       let pdb = _bdbPayloadDb bdb
-      addNewPayload pdb (succ $ view blockHeight $ _parentHeader ph) payload
+      addNewPayload pdb (succ $ _blockHeight $ _parentHeader ph) payload
 
       bhdb <- getBlockHeaderDb cid bdb
       unsafeInsertBlockHeaderDb bhdb bh
