+++ test/Chainweb/Test/Pact/PactSingleChainTest.hs
@@ -60,7 +60,7 @@ import Pact.JSON.Yaml
 
 import Chainweb.BlockCreationTime
 import Chainweb.BlockHash (BlockHash)
-import Chainweb.BlockHeader
+import Chainweb.BlockHeader.Internal
 import Chainweb.BlockHeight (BlockHeight(..))
 import Chainweb.Graph
 import Chainweb.Logger (genericLogger)
@@ -172,11 +172,11 @@ runBlockE :: (HasCallStack) => PactQueue -> TestBlockDb -> TimeSpan Micros -> IO
 runBlockE q bdb timeOffset = do
   T2 (ParentHeader ph) nb <- forSuccess "newBlock" $
         newBlock noMiner q
-  let blockTime = add timeOffset $ _bct $ view blockCreationTime ph
+  let blockTime = add timeOffset $ _bct $ _blockCreationTime ph
   forM_ (chainIds testVersion) $ \c -> do
     let o | c == cid = nb
           | otherwise = emptyPayload
-    addTestBlockDb bdb (succ $ view blockHeight ph) (Nonce 0) (\_ _ -> blockTime) c o
+    addTestBlockDb bdb (succ $ _blockHeight ph) (Nonce 0) (\_ _ -> blockTime) c o
   nextH <- getParentTestBlockDb bdb cid
   validateBlock nextH (payloadWithOutputsToPayloadData nb) q
 
@@ -200,11 +200,11 @@ newBlockAndValidationFailure refIO reqIO = testCase "newBlockAndValidationFailur
 
   T2 (ParentHeader ph) nb <- forSuccess ("newBlockAndValidate" <> ": newblock") $
         newBlock noMiner q
-  let blockTime = add second $ _bct $ view blockCreationTime ph
+  let blockTime = add second $ _bct $ _blockCreationTime ph
   forM_ (chainIds testVersion) $ \c -> do
     let o | c == cid = nb
           | otherwise = emptyPayload
-    addTestBlockDb bdb (succ $ view blockHeight ph) (Nonce 0) (\_ _ -> blockTime) c o
+    addTestBlockDb bdb (succ $ _blockHeight ph) (Nonce 0) (\_ _ -> blockTime) c o
 
   nextH <- getParentTestBlockDb bdb cid
 
@@ -653,8 +653,8 @@ signSender00 = set cbSigners [mkEd25519Signer' sender00 []]
 
 setFromHeader :: BlockHeader -> CmdBuilder -> CmdBuilder
 setFromHeader bh =
-  set cbChainId (view blockChainId bh)
-  . set cbCreationTime (toTxCreationTime $ _bct $ view blockCreationTime bh)
+  set cbChainId (_blockChainId bh)
+  . set cbCreationTime (toTxCreationTime $ _bct $ _blockCreationTime bh)
 
 
 pattern BlockGasLimitError :: forall b. Either PactException b
@@ -801,7 +801,7 @@ moduleNameMempool ns mn = mempty
         fmap V.fromList $ forM (zip txs [0..]) $ \(code,n :: Int) ->
           buildCwCmd ("1" <> sshow n) testVersion $
           signSender00 $
-          set cbCreationTime (toTxCreationTime $ _bct $ view blockCreationTime bh) $
+          set cbCreationTime (toTxCreationTime $ _bct $ _blockCreationTime bh) $
           set cbRPC (mkExec' code) $
           defaultCmd
 
@@ -842,7 +842,7 @@ mempoolCreationTimeTest mpRefIO reqIO = testCase "mempoolCreationTimeTest" $ do
 
     getBlock bh tx valid = do
       let txs = V.singleton tx
-      oks <- valid (view blockHeight bh) (view blockHash bh) txs
+      oks <- valid (_blockHeight bh) (_blockHash bh) txs
       unless (V.and oks) $ throwM $ userError "Insert failed"
       return txs
