+++ test/Chainweb/Test/Pact/TTL.hs
@@ -11,7 +11,7 @@ module Chainweb.Test.Pact.TTL
 ( tests ) where
 
 import Control.Concurrent.MVar
-import Control.Lens (set, view)
+import Control.Lens (set)
 import Control.Monad
 import Control.Monad.Catch
 
@@ -25,7 +25,7 @@ import Test.Tasty.HUnit
 -- internal modules
 
 import Chainweb.BlockCreationTime
-import Chainweb.BlockHeader
+import Chainweb.BlockHeader.Internal
 import Chainweb.BlockHeaderDB hiding (withBlockHeaderDb)
 import Chainweb.BlockHeaderDB.Internal (unsafeInsertBlockHeaderDb)
 import Chainweb.Miner.Pact
@@ -174,7 +174,7 @@ modAt f = modAtTtl f defTtl
 modAtTtl :: (Time Micros -> Time Micros) -> Seconds -> MemPoolAccess
 modAtTtl f (Seconds t) = mempty
     { mpaGetBlock = \_ validate bh hash ph -> do
-        let txTime = toTxCreationTime $ f $ _bct $ view blockCreationTime ph
+        let txTime = toTxCreationTime $ f $ _bct $ _blockCreationTime ph
             tt = TTLSeconds (int t)
         outtxs <- fmap V.singleton $ buildCwCmd (sshow bh) testVer
           $ set cbCreationTime txTime
@@ -221,7 +221,7 @@ doNewBlock ctxIO mempool nonce t = do
     let
         creationTime = BlockCreationTime
             . add (secondsToTimeSpan t) -- 10 seconds
-            . _bct . view blockCreationTime
+            . _bct . _blockCreationTime
             $ _parentHeader parent
         bh = newBlockHeader
             mempty
@@ -243,7 +243,7 @@ doValidateBlock
 doValidateBlock ctxIO header payload = do
     ctx <- ctxIO
     _mv' <- validateBlock header (payloadWithOutputsToPayloadData payload) $ _ctxQueue ctx
-    addNewPayload (_ctxPdb ctx) (view blockHeight header) payload
+    addNewPayload (_ctxPdb ctx) (_blockHeight header) payload
     unsafeInsertBlockHeaderDb (_ctxBdb ctx) header
     -- FIXME FIXME FIXME: do at least some checks?
