+++ test/Chainweb/Test/Pact/Utils.hs
@@ -193,7 +193,7 @@ import Pact.Types.Verifier
 
 -- internal modules
 
-import Chainweb.BlockHeader
+import Chainweb.BlockHeader.Internal
 import Chainweb.BlockHeaderDB hiding (withBlockHeaderDb)
 import Chainweb.BlockHeight
 import Chainweb.ChainId
@@ -785,7 +785,7 @@ runCut v bdb pact genTime noncer miner =
     n <- noncer cid
 
     -- skip this chain if mining fails and retry with the next chain.
-    whenM (addTestBlockDb bdb (succ $ view blockHeight $ _parentHeader ph) n genTime cid pout) $ do
+    whenM (addTestBlockDb bdb (succ $ _blockHeight $ _parentHeader ph) n genTime cid pout) $ do
         h <- getParentTestBlockDb bdb cid
         void $ _webPactValidateBlock pact h (payloadWithOutputsToPayloadData pout)
 
@@ -1050,7 +1050,7 @@ compact logLevel cFlags (SQLiteEnv db _) bh = do
 
 getPWOByHeader :: BlockHeader -> TestBlockDb -> IO PayloadWithOutputs
 getPWOByHeader h (TestBlockDb _ pdb _) =
-  lookupPayloadWithHeight pdb (Just $ view blockHeight h) (view blockPayloadHash h) >>= \case
+  lookupPayloadWithHeight pdb (Just $ _blockHeight h) (_blockPayloadHash h) >>= \case
     Nothing -> throwM $ userError "getPWOByHeader: payload not found"
     Just pwo -> return pwo
