+++ test/Chainweb/Test/TreeDB.hs
@@ -41,7 +41,7 @@ import Test.Tasty.QuickCheck
 
 -- internal modules
 
-import Chainweb.BlockHeader
+import Chainweb.BlockHeader.Internal
 import Chainweb.BlockHeader.Validation
 import Chainweb.Test.Utils
 import Chainweb.Test.Utils.BlockHeader
@@ -234,8 +234,8 @@ entryOrder_prop f (SparseTree t0) = ioProperty . withTreeDb f t $ \db _ -> do
     hs <- entries db Nothing Nothing Nothing Nothing $ P.toList_ . P.map (^. isoBH)
     pure . isJust $ foldlM g S.empty hs
   where
-    g acc h = let acc' = S.insert (view blockHash h) acc
-              in bool Nothing (Just acc') $ isGenesisBlockHeader h || S.member (view blockParent h) acc'
+    g acc h = let acc' = S.insert (_blockHash h) acc
+              in bool Nothing (Just acc') $ isGenesisBlockHeader h || S.member (_blockParent h) acc'
 
     t :: Tree (DbEntry db)
     t = fmap (^. from isoBH) t0
@@ -369,7 +369,7 @@ prop_getBranchIncreasing_parents f (SparseTree t0) = forAll (int <$> choose (0,m
     ioProperty $ withTreeDb f t $ \db _ -> do
         e <- maxEntry db
         branch <- getBranchIncreasing db e i $ \s -> P.toList_ $ P.map (view isoBH) s
-        return $ and $ zipWith (\a b -> view blockHash a == view blockParent b) branch (drop 1 branch)
+        return $ and $ zipWith (\a b -> _blockHash a == _blockParent b) branch (drop 1 branch)
   where
     m = length $ levels t0
     t = fmap (^. from isoBH) t0
