+++ test/Chainweb/Test/Utils.hs
@@ -181,7 +181,7 @@ import UnliftIO.Async
 -- internal modules
 
 import Chainweb.BlockCreationTime
-import Chainweb.BlockHeader
+import Chainweb.BlockHeader.Internal
 import Chainweb.BlockHeaderDB
 import Chainweb.BlockHeaderDB.Internal
 import Chainweb.BlockHeight
@@ -397,8 +397,8 @@ prettyTree :: Tree BlockHeader -> String
 prettyTree = drawTree . fmap f
   where
     f h = printf "%d - %s"
-        (coerce @BlockHeight @Word64 $ view blockHeight h)
-        (take 12 . drop 1 . show $ view blockHash h)
+        (coerce @BlockHeight @Word64 $ _blockHeight h)
+        (take 12 . drop 1 . show $ _blockHash h)
 
 normalizeTree :: Ord a => Tree a -> Tree a
 normalizeTree n@(Node _ []) = n
@@ -438,7 +438,7 @@ genesis v = either (error . sshow) return $ genesisBlockHeaderForChain v 0
 
 forest :: Growth -> BlockHeader -> Gen (Forest BlockHeader)
 forest Randomly h = randomTrunk h
-forest g@(AtMost n) h | n < view blockHeight h = pure []
+forest g@(AtMost n) h | n < _blockHeight h = pure []
                       | otherwise = fixedTrunk g h
 
 fixedTrunk :: Growth -> BlockHeader -> Gen (Forest BlockHeader)
@@ -470,18 +470,18 @@ header p = do
         . newMerkleLog
         $ mkFeatureFlags
             :+: t'
-            :+: view blockHash p
+            :+: _blockHash p
             :+: target
             :+: casKey (testBlockPayloadFromParent (ParentHeader p))
             :+: _chainId p
-            :+: BlockWeight (targetToDifficulty target) + view blockWeight p
-            :+: succ (view blockHeight p)
+            :+: BlockWeight (targetToDifficulty target) + _blockWeight p
+            :+: succ (_blockHeight p)
             :+: _versionCode v
             :+: epochStart (ParentHeader p) mempty t'
             :+: nonce
             :+: MerkleLogBody mempty
    where
-    BlockCreationTime t = view blockCreationTime p
+    BlockCreationTime t = _blockCreationTime p
     target = powTarget (ParentHeader p) mempty t'
     v = _chainwebVersion p
     t' = BlockCreationTime (scaleTimeSpan (10 :: Int) second `add` t)
